refactor: created CpuViewModel and View
