+++ src/Spice86/ViewModels/DebugWindowViewModel.cs
@@ -97,6 +97,9 @@ public partial class DebugWindowViewModel : ViewModelBase, IInternalDebugger, ID
     [ObservableProperty]
     private VideoCardViewModel? _videoCardViewModel;
 
+    [ObservableProperty]
+    private CpuViewModel? _cpuViewModel;
+
     public DebugWindowViewModel(IUIDispatcherTimerFactory iuiDispatcherTimerFactory, IPauseStatus pauseStatus) {
         _pauseStatus = pauseStatus;
         IsPaused = _pauseStatus.IsPaused;
@@ -105,6 +108,7 @@ public partial class DebugWindowViewModel : ViewModelBase, IInternalDebugger, ID
         _iuiDispatcherTimerFactory = iuiDispatcherTimerFactory;
         SoftwareMixerViewModel = new(iuiDispatcherTimerFactory);
         VideoCardViewModel = new();
+        CpuViewModel = new(pauseStatus);
     }
 
     private void OnPauseStatusChanged(object? sender, PropertyChangedEventArgs e) {
@@ -122,12 +126,6 @@ public partial class DebugWindowViewModel : ViewModelBase, IInternalDebugger, ID
         IsLoading = false;
     }
 
-    [ObservableProperty]
-    private StateInfo _state = new();
-
-    [ObservableProperty]
-    private CpuFlagsInfo _flags = new();
-
     public void Visit<T>(T component) where T : IDebuggableComponent {
         if(component is Midi externalMidiDevice) {
             VisitExternalMidiDevice(externalMidiDevice);
@@ -135,9 +133,8 @@ public partial class DebugWindowViewModel : ViewModelBase, IInternalDebugger, ID
         if(component is Cpu cpu) {
             VisitCpu(cpu);
         }
-        if(component is State state) {
-            VisitCpuState(state);
-        }
+        
+        CpuViewModel?.Visit(component);
         
         VideoCardViewModel?.Visit(component);
 
@@ -161,80 +158,6 @@ public partial class DebugWindowViewModel : ViewModelBase, IInternalDebugger, ID
         Midi.LastPortRead = externalMidiDevice.LastPortRead;
         Midi.LastPortWritten = externalMidiDevice.LastPortWritten;
         Midi.LastPortWrittenValue = externalMidiDevice.LastPortWrittenValue;
-    }
-
-    public void VisitCpuState(State state) {
-        if (IsLoading || !IsPaused) {
-            UpdateCpuState(state);
-        }
-
-        if (IsPaused) {
-            State.PropertyChanged -= OnStatePropertyChanged;
-            State.PropertyChanged += OnStatePropertyChanged;
-            Flags.PropertyChanged -= OnStatePropertyChanged;
-            Flags.PropertyChanged += OnStatePropertyChanged;
-        } else {
-            State.PropertyChanged -= OnStatePropertyChanged;
-            Flags.PropertyChanged -= OnStatePropertyChanged;
-        }
-
-        return;
-
-        void OnStatePropertyChanged(object? sender, PropertyChangedEventArgs e) {
-            if (sender is null || e.PropertyName == null || !IsPaused || IsLoading) {
-                return;
-            }
-            PropertyInfo? originalPropertyInfo = state.GetType().GetProperty(e.PropertyName);
-            PropertyInfo? propertyInfo = sender.GetType().GetProperty(e.PropertyName);
-            if (propertyInfo is not null && originalPropertyInfo is not null) {
-                originalPropertyInfo.SetValue(state, propertyInfo.GetValue(sender));
-            }
-        }
-    }
-
-    private void UpdateCpuState(State state) {
-        State.AH = state.AH;
-        State.AL = state.AL;
-        State.AX = state.AX;
-        State.EAX = state.EAX;
-        State.BH = state.BH;
-        State.BL = state.BL;
-        State.BX = state.BX;
-        State.EBX = state.EBX;
-        State.CH = state.CH;
-        State.CL = state.CL;
-        State.CX = state.CX;
-        State.ECX = state.ECX;
-        State.DH = state.DH;
-        State.DL = state.DL;
-        State.DX = state.DX;
-        State.EDX = state.EDX;
-        State.DI = state.DI;
-        State.EDI = state.EDI;
-        State.SI = state.SI;
-        State.ES = state.ES;
-        State.BP = state.BP;
-        State.EBP = state.EBP;
-        State.SP = state.SP;
-        State.ESP = state.ESP;
-        State.CS = state.CS;
-        State.DS = state.DS;
-        State.ES = state.ES;
-        State.FS = state.FS;
-        State.GS = state.GS;
-        State.SS = state.SS;
-        State.IP = state.IP;
-        State.IpPhysicalAddress = state.IpPhysicalAddress;
-        State.StackPhysicalAddress = state.StackPhysicalAddress;
-        State.SegmentOverrideIndex = state.SegmentOverrideIndex;
-        Flags.AuxiliaryFlag = state.AuxiliaryFlag;
-        Flags.CarryFlag = state.CarryFlag;
-        Flags.DirectionFlag = state.DirectionFlag;
-        Flags.InterruptFlag = state.InterruptFlag;
-        Flags.OverflowFlag = state.OverflowFlag;
-        Flags.ParityFlag = state.ParityFlag;
-        Flags.ZeroFlag = state.ZeroFlag;
-        Flags.ContinueZeroFlag = state.ContinueZeroFlagValue;
     }
 
     private bool _needToUpdateDisassembly;
