+++ src/Spice86/Views/CpuView.axaml.cs
@@ -1,0 +1,9 @@
+using Avalonia.Controls;
+
+namespace Spice86.Views;
+
+public partial class CpuView : UserControl {
+    public CpuView() {
+        InitializeComponent();
+    }
+}
